c1

+++ pybricks/pyBricks1/wheelcontroller.py
@@ -129,7 +129,7 @@ class WheelController:
     async def move_wheels_towards_element_then_stop_at_marker():
         Shared.hub().display.icon(Icon.ARROW_UP)
         wheel_controller = WheelController.__object()
-        await wheel_controller.straight(float(80))
+        await wheel_controller.straight(float(70))
 
         while True:
             if await ColorController.detect_white_or_black_mat_color():
